Adding a S3ClientProvider who generates the clients; in this way, S3ClientStore can implement just the caching/storing logic, while the configuration and creation logic of the clients is delegated to the client provider
