+++ src/test/java/software/amazon/nio/spi/s3/S3ClientStoreTest.java
@@ -9,22 +9,13 @@ import junit.framework.TestCase;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.InOrder;
 import org.mockito.Mock;
-import org.mockito.Spy;
 import org.mockito.junit.MockitoJUnitRunner;
-import software.amazon.awssdk.awscore.exception.AwsErrorDetails;
-import software.amazon.awssdk.http.SdkHttpResponse;
 import software.amazon.awssdk.services.s3.S3AsyncClient;
 import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.GetBucketLocationResponse;
-import software.amazon.awssdk.services.s3.model.HeadBucketResponse;
-import software.amazon.awssdk.services.s3.model.S3Exception;
+import org.junit.Rule;
+import org.junit.contrib.java.lang.system.ProvideSystemProperty;
 
-import java.util.NoSuchElementException;
-import java.util.function.Consumer;
-
-import static org.mockito.Mockito.*;
 
 @RunWith(MockitoJUnitRunner.class)
 @SuppressWarnings("unchecked")
@@ -32,11 +23,12 @@ public class S3ClientStoreTest extends TestCase {
 
     S3ClientStore instance;
 
+    @Rule
+    public final ProvideSystemProperty AWS_PROPERTIES
+        = new ProvideSystemProperty("aws.region", "aws-east-1");
+
     @Mock
     S3Client mockClient; //client used to determine bucket location
-
-    @Spy
-    final S3ClientStore spyInstance = S3ClientStore.getInstance();
 
     @Before
     public void setUp() throws Exception {
@@ -51,218 +43,53 @@ public class S3ClientStoreTest extends TestCase {
 
     @Test
     public void testGetClientForNullBucketName() {
-        assertEquals(S3ClientStore.DEFAULT_CLIENT, instance.getClientForBucketName(null));
+        assertEquals(instance.provider.universalClient(), instance.getClientForBucketName(null));
     }
 
     @Test
     public void testGetAsyncClientForNullBucketName() {
-        assertEquals(S3ClientStore.DEFAULT_ASYNC_CLIENT, instance.getAsyncClientForBucketName(null));
+        assertEquals(instance.provider.universalClient(true), instance.getAsyncClientForBucketName(null));
     }
 
     @Test
     public void testGetClientForEmptyBucketName() {
-        assertEquals(S3ClientStore.DEFAULT_CLIENT, instance.getClientForBucketName(""));
-        assertEquals(S3ClientStore.DEFAULT_CLIENT, instance.getClientForBucketName(" "));
+        assertEquals(instance.provider.universalClient(), instance.getClientForBucketName(""));
+        assertEquals(instance.provider.universalClient(), instance.getClientForBucketName(" "));
     }
 
     @Test
     public void testGetAsyncClientForEmptyBucketName() {
-        assertEquals(S3ClientStore.DEFAULT_ASYNC_CLIENT, instance.getAsyncClientForBucketName(""));
-        assertEquals(S3ClientStore.DEFAULT_ASYNC_CLIENT, instance.getAsyncClientForBucketName(" "));
-    }
-
-    @Test
-    public void testGenerateClientWithNoErrors() {
-        when(mockClient.getBucketLocation(any(Consumer.class)))
-                .thenReturn(GetBucketLocationResponse.builder().locationConstraint("us-west-2").build());
-        final S3Client s3Client = instance.generateClient("test-bucket", mockClient);
-        assertNotNull(s3Client);
-
-    }
-
-    @Test
-    public void testGenerateAsyncClientWithNoErrors() {
-        when(mockClient.getBucketLocation(any(Consumer.class)))
-                .thenReturn(GetBucketLocationResponse.builder().locationConstraint("us-west-2").build());
-        final S3AsyncClient s3Client = instance.generateAsyncClient("test-bucket", mockClient);
-        assertNotNull(s3Client);
-    }
-
-    @Test
-    public void testGenerateClientWith403Response() {
-        // when you get a forbidden response from getBucketLocation
-        when(mockClient.getBucketLocation(any(Consumer.class))).thenThrow(
-                S3Exception.builder().statusCode(403).build()
-        );
-        // you should fall back to a head bucket attempt
-        when(mockClient.headBucket(any(Consumer.class)))
-                .thenReturn((HeadBucketResponse) HeadBucketResponse.builder()
-                        .sdkHttpResponse(SdkHttpResponse.builder()
-                                .putHeader("x-amz-bucket-region", "us-west-2")
-                                .build())
-                        .build());
-
-        // which should get you a client
-        final S3Client s3Client = instance.generateClient("test-bucket", mockClient);
-        assertNotNull(s3Client);
-
-        final InOrder inOrder = inOrder(mockClient);
-        inOrder.verify(mockClient).getBucketLocation(any(Consumer.class));
-        inOrder.verify(mockClient).headBucket(any(Consumer.class));
-        inOrder.verifyNoMoreInteractions();
-    }
-
-    @Test
-    public void testGenerateAsyncClientWith403Response() {
-        // when you get a forbidden response from getBucketLocation
-        when(mockClient.getBucketLocation(any(Consumer.class))).thenThrow(
-                S3Exception.builder().statusCode(403).build()
-        );
-        // you should fall back to a head bucket attempt
-        when(mockClient.headBucket(any(Consumer.class)))
-                .thenReturn((HeadBucketResponse) HeadBucketResponse.builder()
-                        .sdkHttpResponse(SdkHttpResponse.builder()
-                                .putHeader("x-amz-bucket-region", "us-west-2")
-                                .build())
-                        .build());
-
-        // which should get you a client
-        final S3AsyncClient s3Client = instance.generateAsyncClient("test-bucket", mockClient);
-        assertNotNull(s3Client);
-
-        final InOrder inOrder = inOrder(mockClient);
-        inOrder.verify(mockClient).getBucketLocation(any(Consumer.class));
-        inOrder.verify(mockClient).headBucket(any(Consumer.class));
-        inOrder.verifyNoMoreInteractions();
-    }
-
-    @Test
-    public void testGenerateAsyncClientWith403Then301Responses(){
-        // when you get a forbidden response from getBucketLocation
-        when(mockClient.getBucketLocation(any(Consumer.class))).thenThrow(
-                S3Exception.builder().statusCode(403).build()
-        );
-        // and you get a 301 response on headBucket
-        when(mockClient.headBucket(any(Consumer.class))).thenThrow(
-                S3Exception.builder()
-                        .statusCode(301)
-                        .awsErrorDetails(AwsErrorDetails.builder()
-                                .sdkHttpResponse(SdkHttpResponse.builder()
-                                        .putHeader("x-amz-bucket-region", "us-west-2")
-                                        .build())
-                                .build())
-                        .build()
-        );
-
-        // then you should be able to get a client as long as the error response header contains the region
-        final S3AsyncClient s3Client = instance.generateAsyncClient("test-bucket", mockClient);
-        assertNotNull(s3Client);
-
-        final InOrder inOrder = inOrder(mockClient);
-        inOrder.verify(mockClient).getBucketLocation(any(Consumer.class));
-        inOrder.verify(mockClient).headBucket(any(Consumer.class));
-        inOrder.verifyNoMoreInteractions();
-    }
-
-    @Test
-    public void testGenerateClientWith403Then301ResponsesNoHeader(){
-        // when you get a forbidden response from getBucketLocation
-        when(mockClient.getBucketLocation(any(Consumer.class))).thenThrow(
-                S3Exception.builder().statusCode(403).build()
-        );
-        // and you get a 301 response on headBucket but no header for region
-        when(mockClient.headBucket(any(Consumer.class))).thenThrow(
-                S3Exception.builder()
-                        .statusCode(301)
-                        .awsErrorDetails(AwsErrorDetails.builder()
-                                .sdkHttpResponse(SdkHttpResponse.builder()
-                                        .build())
-                                .build())
-                        .build()
-        );
-
-        // then you should get a NoSuchElement exception when you try to get the header
-        try {
-            instance.generateClient("test-bucket", mockClient);
-        } catch (Exception e) {
-            assertEquals(NoSuchElementException.class, e.getClass());
-        }
-
-        final InOrder inOrder = inOrder(mockClient);
-        inOrder.verify(mockClient).getBucketLocation(any(Consumer.class));
-        inOrder.verify(mockClient).headBucket(any(Consumer.class));
-        inOrder.verifyNoMoreInteractions();
-    }
-
-
-    @Test
-    public void testGenerateAsyncClientWith403Then301ResponsesNoHeader(){
-        // when you get a forbidden response from getBucketLocation
-        when(mockClient.getBucketLocation(any(Consumer.class))).thenThrow(
-                S3Exception.builder().statusCode(403).build()
-        );
-        // and you get a 301 response on headBucket but no header for region
-        when(mockClient.headBucket(any(Consumer.class))).thenThrow(
-                S3Exception.builder()
-                        .statusCode(301)
-                        .awsErrorDetails(AwsErrorDetails.builder()
-                                .sdkHttpResponse(SdkHttpResponse.builder()
-                                        .build())
-                                .build())
-                        .build()
-        );
-
-        // then you should get a NoSuchElement exception when you try to get the header
-        try {
-            instance.generateAsyncClient("test-bucket", mockClient);
-        } catch (Exception e) {
-            assertEquals(NoSuchElementException.class, e.getClass());
-        }
-
-        final InOrder inOrder = inOrder(mockClient);
-        inOrder.verify(mockClient).getBucketLocation(any(Consumer.class));
-        inOrder.verify(mockClient).headBucket(any(Consumer.class));
-        inOrder.verifyNoMoreInteractions();
+        assertEquals(instance.provider.universalClient(true), instance.getAsyncClientForBucketName(""));
+        assertEquals(instance.provider.universalClient(true), instance.getAsyncClientForBucketName(" "));
     }
 
     @Test
     public void testCaching() {
-        S3Client client = S3Client.create();
-        doReturn(client).when(spyInstance).generateClient("test-bucket");
+        instance.provider = new S3ClientProvider() {
+            @Override
+            protected S3Client generateClient(String bucketName) {
+                return S3Client.create();
+            }
+        };
 
-        final S3Client client1 = spyInstance.getClientForBucketName("test-bucket");
-        verify(spyInstance).generateClient("test-bucket");
-        assertSame(client1, client);
-
-        S3Client differentClient = S3Client.create();
-        assertNotSame(client, differentClient);
-
-        lenient().doReturn(differentClient).when(spyInstance).generateClient("test-bucket");
-        final S3Client client2 = spyInstance.getClientForBucketName("test-bucket");
-        // same instance because second is cached.
-        assertSame(client1, client2);
-        assertSame(client2, client);
-        assertNotSame(client2, differentClient);
+        final S3Client client1 = instance.getClientForBucketName("test-bucket1");
+        assertSame(client1, instance.getClientForBucketName("test-bucket1"));
+        assertNotSame(client1, instance.getClientForBucketName("test-bucket2"));
     }
 
     @Test
     public void testAsyncCaching() {
-        S3AsyncClient client = S3AsyncClient.create();
-        doReturn(client).when(spyInstance).generateAsyncClient("test-bucket");
+        instance.provider = new S3ClientProvider() {
+            @Override
+            protected S3AsyncClient generateAsyncClient(String bucketName) {
+                return S3AsyncClient.create();
+            }
+        };
 
-        final S3AsyncClient client1 = spyInstance.getAsyncClientForBucketName("test-bucket");
-        verify(spyInstance).generateAsyncClient("test-bucket");
-        assertSame(client1, client);
-
-        S3AsyncClient differentClient = S3AsyncClient.create();
-        assertNotSame(client, differentClient);
-
-        lenient().doReturn(differentClient).when(spyInstance).generateAsyncClient("test-bucket");
-        final S3AsyncClient client2 = spyInstance.getAsyncClientForBucketName("test-bucket");
-        // same instance because second is cached.
-        assertSame(client1, client2);
-        assertSame(client2, client);
-        assertNotSame(client2, differentClient);
+        final S3AsyncClient client1 = instance.getAsyncClientForBucketName("test-bucket1");
+        assertSame(client1, instance.getAsyncClientForBucketName("test-bucket1"));
+        assertNotSame(client1, instance.getAsyncClientForBucketName("test-bucket2"));
     }
+
 }
